backend/ocr-api/requirements.txt
fastapi
uvicorn
opencv-python
pillow
pytesseract
ftfy
python-multipart
requests